Renamed core.process to core.tools and moved things into it.
