+++ holopy/core/__init__.py
@@ -39,9 +39,7 @@ Main use cases are
 """
 
 from .marray import (Marray, Image, Schema, ImageSchema, Volume,
-                    VolumeSchema, VectorGrid, VectorGridSchema,
-                    subimage)
+                    VolumeSchema, VectorGrid, VectorGridSchema)
 from .metadata import Optics, Grid, Angles, UnevenGrid, interpret_args
 from .io import load, load_image, save, save_image
-from . import process
-from . import helpers
+from . import tools
