+++ holopy/core/metadata.py
@@ -26,7 +26,7 @@ Classes for defining metadata about experimental or calculated results.
 import numpy as np
 from warnings import warn
 import copy
-from .helpers import _ensure_pair, _ensure_array
+from .tools.helpers import _ensure_pair, _ensure_array
 from .holopy_object import HoloPyObject
