+++ holopy/core/tests/test_marray.py
@@ -20,8 +20,8 @@ import numpy as np
 from nose.tools import nottest
 from numpy.testing import assert_equal, assert_allclose, assert_raises
 from ..marray import (ImageSchema, VectorGrid, make_vector_schema,
-                      Image, zeros_like, subimage, resize, Volume,
-                      Schema, squeeze)
+                      Image, zeros_like, Volume,
+                      Schema)
 from .. import Optics
 from .common import assert_obj_close, get_example_data
 from ..errors import UnspecifiedPosition
@@ -83,39 +83,6 @@ def test_resample():
     assert_obj_close(i, Image(np.arange(16).reshape((4,4)),
                               spacing = 1), context='image')
 
-def test_subimage():
-    i = np.zeros((10,10))
-    s = subimage(i, (5,5), 2)
-    assert s.shape == (2,2)
-
-    i2 = Image(i, 1)
-    s2 = subimage(i2, (5, 5), 2)
-
-    assert_equal(s2.positions, subimage(i2.positions, (5,5), 2))
-
-    assert_raises(IndexError, subimage, i, (2,2), 10)
-
-def test_subimage_floats():
-    i = Image(np.zeros((100, 100)), .1)
-    s1 = subimage(i, (5.2,5.6), 2)
-    s2 = subimage(i, (5,6), 2)
-    assert_equal(s1, s2)
-
-def test_resize():
-    i = Image(np.zeros((100, 100)), .1)
-    r = resize(i, (5, 5), (8, 8))
-    assert_equal(r.center, (5, 5, 0))
-    assert_equal(r.extent[:2], (8, 8))
-
-    v = Volume(np.zeros((10, 10, 10)), .1)
-    r = resize(v, extent =  (.8, .8, .8))
-    assert_equal(r.center, (.5, .5, .5))
-    assert_equal(r.extent, (.8, .8, .8))
-
-    r = resize(v, spacing = (.2, .2, .2))
-    assert_equal(r.spacing, (.2, .2, .2))
-    assert_equal(r.shape, (5, 5, 5))
-
 def test_positions_theta_phi():
     theta = np.linspace(0, np.pi)
     phi = np.linspace(0, 2*np.pi)
@@ -130,14 +97,6 @@ def test_positions_shape():
     schema = ImageSchema(shape = 105, spacing = 0.041)
     pos = schema.positions.r_theta_phi([0,0,0])
     assert_equal(pos.shape[0], schema.shape[0]**2)
-
-def test_squeeze():
-    v = Volume(np.ones((10, 1, 10)), spacing = (1, 2, 3),
-                       optics = Optics(.66, 1, (1, 0)))
-    s = squeeze(v)
-    assert_equal(s.shape, (10, 10))
-    assert_equal(s.optics, v.optics)
-    assert_equal(s.spacing, (1, 3))
 
 def test_metadata_transfer():
     s = Schema(100)
