+++ holopy/core/tools/fourier.py
@@ -22,10 +22,7 @@ Misc utility functions to make coding more convenient
 """
 
 
-import scipy.fftpack as fftpack
-from numpy import sin, cos
-import numpy as np
-from .marray import arr_like, Marray
+from scipy import fftpack
 
 
 def fft(a, overwrite=False, shift=True):
@@ -67,8 +64,6 @@ def fft(a, overwrite=False, shift=True):
                                     axes=[0,1])
         else:
             res = fftpack.fft2(a, axes=[0, 1], overwrite_x=overwrite)
-    if isinstance(a, Marray):
-        res = arr_like(res, a)
     return res
 
 
@@ -110,76 +105,5 @@ def ifft(a, overwrite=False, shift=True):
                                  overwrite_x=overwrite)
         else:
             res = fftpack.ifft2(a, overwrite_x=overwrite)
-    if isinstance(a, Marray):
-        res = arr_like(res, a)
     return res
 
-def rotate_points(points, theta, phi, psi):
-    points = np.array(points)
-    rot = rotation_matrix(theta, phi, psi)
-    if points.ndim == 1:
-        return np.dot(rot, points)
-    return np.array([np.dot(rot, c) for c in points])
-
-def rotation_matrix(alpha, beta, gamma, radians = True):
-    """
-    Return a 3D rotation matrix
-
-    Parameters
-    ----------
-    alpha, beta, gamma: float
-        Euler rotation angles in z, y, z convention
-    radians: boolean
-        Default True; treats input angles in radians
-
-    Returns
-    -------
-    rot: array(3,3)
-        Rotation matrix. To rotate a column vector x, use np.dot(rot, x.) 
-
-    Notes
-    -----
-    The Euler angles rotate a vector (in the active picture) by alpha
-    clockwise about the fixed lab z axis, beta clockwise about
-    the lab y axis, and by gamma about the lab z axis.  Clockwise is
-    defined as viewed from the origin, looking in the positive direction
-    along an axis.  
-
-    This breaks compatability with previous conventions, which were adopted for 
-    compatability with the passive picture used by SCSMFO.
-
-    """
-    if not radians:
-        alpha *= np.pi/180.
-        beta *= np.pi/180.
-        gamma *= np.pi/180.
-
-    ca = cos(alpha)
-    sa = sin(alpha)
-    cb = cos(beta)
-    sb = sin(beta)
-    cg = cos(gamma)
-    sg = sin(gamma)
-
-    return np.array([ca*cb*cg - sa*sg, -sa*cb*cg - ca*sg, sb*cg,
-                     ca*cb*sg + sa*cg, -sa*cb*sg + ca*cg, sb*sg,
-                     -ca*sb, sa*sb, cb]).reshape((3,3)) # row major
-
-def cartesian_distance(p1, p2):
-    """
-    Return the Cartesian distance between points p1 and p2.
-
-    Parameters
-    ----------
-    p1, p2: array or list
-        Coordinates of point 1 and point 2 in N-dimensions
-
-    Returns
-    -------
-    dist: float64
-        Cartesian distance between points p1 and p2 
-
-    """
-    p1 = np.array(p1)
-    p2 = np.array(p2)
-    return np.sqrt(np.sum((p1-p2)**2))
