+++ holopy/fitting/fit_series.py
@@ -29,9 +29,9 @@ import warnings
 import os
 import types
 import numpy as np
-from holopy.core.process import normalize
-from holopy.core import subimage, Image
-from holopy.core.helpers import mkdir_p
+from holopy.core.tools import normalize, subimage
+from holopy.core import Image
+from holopy.core.tools.helpers import mkdir_p
 from holopy.core.io import load, save
 from holopy.fitting import fit
