+++ holopy/inference/noise_model.py
@@ -20,7 +20,7 @@ from holopy.fitting.model import BaseModel
 from holopy.fitting.parameter import Parameter
 from holopy.core.holopy_object import HoloPyObject
 from holopy.core import Marray
-from holopy.core.helpers import dict_without
+from holopy.core.tools.helpers import dict_without
 from holopy.scattering.errors import MultisphereExpansionNaN, MultisphereFieldNaN, ConvergenceFailureMultisphere, ScattererDefinitionError
 
 import numpy as np
