+++ holopy/scattering/scatterer/composite.py
@@ -31,7 +31,7 @@ import numpy as np
 
 
 from . import Scatterer
-from ...core.math import rotate_points
+from ...core.tools.math import rotate_points
 
 class Scatterers(Scatterer):
     '''
